+++ src/main/java/dandi/dandi/comment/application/port/in/CommentQueryServicePort.java
@@ -1,0 +1,8 @@
+package dandi.dandi.comment.application.port.in;
+
+import org.springframework.data.domain.Pageable;
+
+public interface CommentQueryServicePort {
+
+    CommentResponses getComments(Long memberId, Long postId, Pageable pageable);
+}
