+++ src/main/java/dandi/dandi/comment/application/port/in/CommentReportUseCaseServicePort.java
@@ -1,0 +1,6 @@
+package dandi.dandi.comment.application.port.in;
+
+public interface CommentReportUseCaseServicePort {
+
+    void reportComment(Long memberId, Long commentId);
+}
