--- src/main/java/dandi/dandi/comment/application/port/in/CommentUseCase.java
@@ -1,14 +1,0 @@
-package dandi.dandi.comment.application.port.in;
-
-import org.springframework.data.domain.Pageable;
-
-public interface CommentUseCase {
-
-    void registerComment(Long memberId, Long postId, CommentRegisterCommand commentRegisterCommand);
-
-    CommentResponses getComments(Long memberId, Long postId, Pageable pageable);
-
-    void deleteComment(Long memberId, Long commentId);
-
-    void reportComment(Long memberId, Long commentId);
-}
