+++ src/main/java/dandi/dandi/comment/application/port/in/CommentUseCaseServicePort.java
@@ -1,0 +1,8 @@
+package dandi.dandi.comment.application.port.in;
+
+public interface CommentUseCaseServicePort {
+
+    void registerComment(Long memberId, Long postId, CommentRegisterCommand commentRegisterCommand);
+
+    void deleteComment(Long memberId, Long commentId);
+}
